Create Common.js
This is my common js file. It is a collection of prototypes from Mozilla, and a couple of helpers from StackOverflow.

If you are going to comment about my license being wrong for this, then please note that I am still compiling all of the licenses that I need, and will add them to the code. The good news is that off of the licenses that I have pulled this from are from open source lic. sites.